moved the logic of exporting to  a new util file
